requirements.txt
# Google Sheets API
google-api-python-client==2.108.0
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Chrome automation
selenium==4.15.2
webdriver-manager==4.0.1
undetected-chromedriver==3.5.4

# Data handling
pandas==2.1.4
openpyxl==3.1.2

# Utility
python-dotenv==1.0.0
requests==2.31.0

# AI API
google-generativeai==0.8.3 